+++ console/src/components/Services/Data/TableRelationships/RemoteRelationships/SchemaExplorer.scss
@@ -29,3 +29,10 @@
   height: 15px;
   background-color: white;
 }
+
+.scalarColumnSelect {
+  margin-left:  5px;
+  border-radius: 4px;
+  background-color: white; 
+  border: 1px solid #ccc;
+}
