+++ console/src/helpers/versionUtils.js
@@ -1,11 +1,13 @@
 const semver = require('semver');
 
 export const FT_JWT_ANALYZER = 'JWTAnalyzer';
+export const FT_REMOTE_RELATIONSHIPS = 'RemoteRelationships';
 
 // list of feature launch versions
 const featureLaunchVersions = {
   // feature: 'v1.0.0'
   [FT_JWT_ANALYZER]: 'v1.0.0-beta.3',
+  [FT_REMOTE_RELATIONSHIPS]: 'v1.0.0-beta.3',
 };
 
 export const getFeaturesCompatibility = serverVersion => {
